Refactor: extract adding event listeners into separate functions and move them into their respective packages
